CI: Distribute kafka_storage tests in 4 modules
